refactor: move the Step Functions CFN factory
The factory for AWS::StepFunctions::StateMachine now sits at
src/service/stepfunctions/cfn/, where every other service keeps its
CloudFormation factory, and simulated Step Functions hands it over
through cfnResourceFactory() like the rest of them. It was put under
src/service/cloudformation/resource/cfn/stepfunctions/ to stay clear of
the interpreter work on #918, which has landed. The value adapter stays
where the other services keep theirs. This is a rename, an import
rewrite and one edit to the service factory map.
