+++ src/service/cloudformation/resource/resolve/service/sim-cfn-service-factories.ts
@@ -1,7 +1,6 @@
 import type { SimAwsAccountRegionContainer } from "../../../../aws/sim-aws-account-region-scope.js";
 import type { SimCfnServiceResourceFactory } from "../../factory/sim-cfn-resource-factory.type.js";
 import { SimCfnCfnResourceFactory } from "../../factory/sim-cfn-cfn-resource-factory.js";
-import { SimStepFunctionsCfnResourceFactory } from "../../cfn/stepfunctions/sim-step-functions-cfn-resource-factory.js";
 
 /**
  * How one simulated service hands over its CloudFormation Resource factory,
@@ -168,9 +167,7 @@ export const simCfnServiceResourceFactories: ReadonlyMap<
   [
     "StepFunctions",
     (scopedAws): SimCfnServiceResourceFactory =>
-      new SimStepFunctionsCfnResourceFactory({
-        stepFunctions: scopedAws.stepFunctions(),
-      }),
+      scopedAws.stepFunctions().cfnResourceFactory(),
   ],
   [
     "WAFv2",
