+++ src/service/stepfunctions/cfn/sim-cfn-state-machine-property-refusals.iso.test.ts
@@ -5,10 +5,10 @@ import {
 } from "@kensio/smartass";
 import { describe, it } from "vitest";
 
-import { SimAws } from "../../../../aws/sim-aws.js";
-import { SimStateMachine } from "../../../../stepfunctions/machine/sim-state-machine.js";
-import type { SimCfnTemplateValueRecord } from "../../../template/value/sim-cfn-template-value.js";
-import { SimStateMachineCfn } from "./sim-state-machine-cfn.js";
+import { SimAws } from "../../aws/sim-aws.js";
+import { SimStateMachine } from "../machine/sim-state-machine.js";
+import type { SimCfnTemplateValueRecord } from "../../cloudformation/template/value/sim-cfn-template-value.js";
+import { SimStateMachineCfn } from "../../cloudformation/resource/cfn/stepfunctions/sim-state-machine-cfn.js";
 
 describe("What an AWS::StepFunctions::StateMachine Resource shape refuses", () => {
   const roleArn = "arn:aws:iam::123456789012:role/EnrolmentWorkflowRole";
