+++ src/service/stepfunctions/cfn/sim-cfn-step-functions-resource-deleter.ts
@@ -1,7 +1,7 @@
-import { assertDefined } from "../../../../../util/type-guard/defined.js";
-import type { SimStateMachine } from "../../../../stepfunctions/machine/sim-state-machine.js";
-import type { SimStepFunctions } from "../../../../stepfunctions/sim-step-functions.js";
-import type { SimCfnResource } from "../../sim-cfn-resource.js";
+import { assertDefined } from "../../../util/type-guard/defined.js";
+import type { SimStateMachine } from "../machine/sim-state-machine.js";
+import type { SimStepFunctions } from "../sim-step-functions.js";
+import type { SimCfnResource } from "../../cloudformation/resource/sim-cfn-resource.js";
 import { stateMachineResourceTypeName } from "./sim-cfn-step-functions-resource-types.js";
 
 interface SimCfnStepFunctionsResourceDeleterProperties {
