+++ src/service/stepfunctions/cfn/sim-cfn-step-functions-resource-types.ts
