+++ src/service/stepfunctions/cfn/sim-step-functions-cfn-resource-factory.ts
@@ -1,11 +1,11 @@
-import type { SimStepFunctions } from "../../../../stepfunctions/sim-step-functions.js";
-import type { SimCfnServiceResourceFactory } from "../../factory/sim-cfn-resource-factory.type.js";
+import type { SimStepFunctions } from "../sim-step-functions.js";
+import type { SimCfnServiceResourceFactory } from "../../cloudformation/resource/factory/sim-cfn-resource-factory.type.js";
 import type {
   SimCfnResource,
   SimCloudFormationResourceCreateContext,
   SimCloudFormationResourceDeleteContext,
-} from "../../sim-cfn-resource.js";
-import { SimCfnStateMachineCreator } from "./sim-cfn-state-machine-creator.js";
+} from "../../cloudformation/resource/sim-cfn-resource.js";
+import { SimCfnStateMachineCreator } from "./state-machine/sim-cfn-state-machine-creator.js";
 import { SimCfnStepFunctionsResourceDeleter } from "./sim-cfn-step-functions-resource-deleter.js";
 import { stateMachineResourceTypeName } from "./sim-cfn-step-functions-resource-types.js";
