+++ src/service/stepfunctions/cfn/state-machine/sim-cfn-state-machine-property-names.ts
